user_register/requirements.txt
alembic==1.9.4
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
